// src/app/shared-module/enum/language.enum.ts
/**
 * 国际化枚举
 */

export enum LanguageEnum {
  // 设施
  facility = 'facility',
  // 公共国际化
  common = 'common',
  // 告警国际化
  alarm = 'alarm',
  // 工单
  workOrder = 'workOrder',
  // 系统设置
  systemSetting = 'systemSetting',
  // 首页
  index = 'index',
  // 资产（新增的）
  assets = 'assets',
  // form
  form = 'form',
  // 巡检
  inspection = 'inspection',
  // 表格配置
  online = 'online',
  // 应用系统
  application = 'application',
  fault = 'fault',
  // 日志
  log = 'log',
  // 用户国际化
  user = 'user',
  role = 'role',
  // 单位国际化
  unit = 'unit',
  // 大屏
  bigScreen = 'bigScreen',
  // 导航菜单
  navigation = 'navigation',
  // 下载模块
  download = 'download',
  // 统计
  statistical = 'statistical',
  // 设施授权
  facilityAuth = 'facilityAuth',
  // 产品管理国际化
  product = 'product',
  // 租户管理
  tenant = 'tenant',
  // 表格
  table = 'table',
  // 排班管理
  schedule = 'schedule',
  // 规划项目管理
  planProject = 'planProject',
  // 库存管理
  storage = 'storage',
  // 供应商管理
  supplier = 'supplier',
  // 能耗
  energy= 'energy'
}